Ability System
The ability system completely overhauled, now it can scale more easily.
Added a new level "The Dark Forest" and added some decoration for it too.
And finally some minor changes to the enemies and the minimap, now a red dot appears to mark where the enemies are on the map.