Major Changes
Added more dungeon rooms for the forest level.
Added more music.
Added two more weapons, the shotgun and showed shotgun.
Now the bullets of the player's weapons have a sound effect when colliding with the environment.
#Audio System
Now the audio system uses Events to configure the volume settings instead of relying on script references.

And More.